requirements.txt
numpy>=1.24
pandas>=2.0
jupyter>=1.0
jupyterlab>=4.0
matplotlib>=3.7
seaborn>=0.12
plotly>=5.15
folium>=0.14
scipy>=1.11
statsmodels>=0.14
scikit-learn>=1.3
xgboost>=2.0
prophet>=1.1
nbconvert>=7.0
google-cloud-bigquery>=3.11
pandas-gbq>=0.19
db-dtypes>=1.1
pyarrow>=14.0
